Java port development abandoned
Due to a combination of insufficient motivation and benefit, I've
decided to indefinitely stop development of a Java GUI port of this
application. In my eyes, the functionality of the script is not
advanced enough to warrant a GUI application and all the additional
programming associated with it. Any further advancements or improvements
will be made to the python script.